+++ samples/lib/widgets/calendar.dart
@@ -57,7 +57,7 @@ class UnselectableCalendarPage extends SampleScaffold {
   Widget child(BuildContext context) => FCalendar(
         controller: FCalendarMultiValueController(
           initialSelections: {DateTime.utc(2024, 7, 17), DateTime.utc(2024, 7, 20)},
-          canSelect: (date) => !{DateTime.utc(2024, 7, 18), DateTime.utc(2024, 7, 19)}.contains(date),
+          selectable: (date) => !{DateTime.utc(2024, 7, 18), DateTime.utc(2024, 7, 19)}.contains(date),
         ),
         start: DateTime.utc(2000),
         today: DateTime.utc(2024, 7, 15),